@@ -17,24 +17,24 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "11.1.1",
-    "@angular/cdk": "11.1.1",
-    "@angular/common": "11.1.1",
-    "@angular/compiler": "11.1.1",
-    "@angular/core": "11.1.1",
-    "@angular/forms": "11.1.1",
-    "@angular/localize": "11.1.1",
-    "@angular/material": "11.1.1",
-    "@angular/platform-browser": "11.1.1",
-    "@angular/platform-browser-dynamic": "11.1.1",
-    "@angular/router": "11.1.1",
+    "@angular/animations": "11.2.0",
+    "@angular/cdk": "11.2.0",
+    "@angular/common": "11.2.0",
+    "@angular/compiler": "11.2.0",
+    "@angular/core": "11.2.0",
+    "@angular/forms": "11.2.0",
+    "@angular/localize": "11.2.0",
+    "@angular/material": "11.2.0",
+    "@angular/platform-browser": "11.2.0",
+    "@angular/platform-browser-dynamic": "11.2.0",
+    "@angular/router": "11.2.0",
     "@auth0/angular-jwt": "5.0.2",
-    "@ngrx/component": "10.1.2",
-    "@ngrx/effects": "10.1.2",
-    "@ngrx/entity": "10.1.2",
-    "@ngrx/router-store": "10.1.2",
-    "@ngrx/store": "10.1.2",
-    "@ngrx/store-devtools": "10.1.2",
+    "@ngrx/component": "11.0.0",
+    "@ngrx/effects": "11.0.0",
+    "@ngrx/entity": "11.0.0",
+    "@ngrx/router-store": "11.0.0",
+    "@ngrx/store": "11.0.0",
+    "@ngrx/store-devtools": "11.0.0",
     "@swimlane/ngx-charts": "17.0.0",
     "bootstrap": "4.6.0",
     "bootstrap-icons": "1.3.0",
@@ -42,35 +42,33 @@
     "filepond-plugin-file-validate-type": "1.2.5",
     "filepond-plugin-image-crop": "2.0.4",
     "filepond-plugin-image-exif-orientation": "1.0.9",
-    "filepond-plugin-image-preview": "4.6.4",
+    "filepond-plugin-image-preview": "4.6.5",
     "filepond-plugin-image-resize": "2.0.7",
-    "filepond-plugin-image-transform": "3.7.4",
-    "lodash": "4.17.20",
+    "filepond-plugin-image-transform": "3.7.5",
     "ngx-filepond": "6.0.0",
     "normalizr": "3.6.1",
     "rxjs": "6.6.3",
-    "socket.io-client": "3.1.0",
+    "socket.io-client": "3.1.1",
     "tslib": "2.1.0",
-    "ua-parser-js": "0.7.23",
+    "ua-parser-js": "0.7.24",
     "uuid": "8.3.2",
     "zone.js": "0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1101.2",
-    "@angular/cli": "11.1.2",
-    "@angular/compiler-cli": "11.1.1",
-    "@angular/language-service": "11.1.1",
-    "@ngrx/schematics": "10.1.2",
+    "@angular-devkit/build-angular": "0.1102.0",
+    "@angular/cli": "11.2.0",
+    "@angular/compiler-cli": "11.2.0",
+    "@angular/language-service": "11.2.0",
+    "@ngrx/schematics": "11.0.0",
     "@types/jasmine": "3.6.3",
     "@types/jasminewd2": "2.0.8",
-    "@types/lodash": "4.14.168",
-    "@types/node": "13.13.30",
+    "@types/node": "13.13.41",
     "@types/ua-parser-js": "0.7.35",
     "@types/uuid": "8.3.0",
     "codelyzer": "6.0.1",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
-    "karma": "6.0.3",
+    "karma": "6.1.1",
     "karma-chrome-launcher": "3.1.0",
     "karma-coverage": "2.0.3",
     "karma-jasmine": "4.0.1",
@@ -80,7 +78,7 @@
     "ts-node": "9.1.1",
     "tslint": "6.1.3",
     "tslint-config-prettier": "1.18.0",
-    "typescript": "4.1.3",
+    "typescript": "4.1.5",
     "webpack-bundle-analyzer": "4.4.0"
   }
 }
