@@ -1,19 +1,20 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component } from '@angular/core';
-import { takeUntil } from 'rxjs/operators';
+import { first } from 'rxjs/operators';
 import { Observable } from 'rxjs';
 
 import { select, Store } from '@ngrx/store';
 
 import { AppState } from '@/core/store';
-import { User } from '@/core/modules/user-api/shared';
 import { Role, Privilege } from '@/core/modules/role-api/shared';
 import { selectRolesLoading } from '@/core/modules/role-api/store/role';
+import { selectUsersById } from '@/core/modules/user-api/store';
 import { TemplatesViewDetailsComponent } from '@/shared/components';
 import { MAT_BUTTON, COLUMN_KEYS, COLUMN_LABELS, CONSTANTS, BS_ICONS, FORMCONSTANTS } from '@/shared/constants';
 import { DialogService } from '@/shared/services';
-import { DialogDataModel, DialogResultModel, DialogType, DialogWithTwoButtonModel } from '@/shared/models';
+import { DialogDataModel, IDialogResult, DialogType, DialogWithTwoButtonModel } from '@/shared/models';
 import { UsersDialogComponent } from '@/modules/user-management/components';
 import { PrivilegesDialogComponent } from '@/modules/role-management/components/privileges/dialog/privileges-dialog.component';
+import { resetSelectionPopup, prepareSelectionPopup } from '@/modules/user-management/store';
 
 @Component({
   selector: 'templates-role-view',
@@ -50,31 +51,32 @@ export class TemplatesRoleViewComponent extends TemplatesViewDetailsComponent<Ro
     super(store$, dialogService);
   }
 
-  addParticipantDialog(): void {
+  addUserDialog(): void {
     const dialogDataModel: DialogDataModel<DialogWithTwoButtonModel> = {
-      dialogModel: new DialogWithTwoButtonModel(CONSTANTS.TEXTS_SELECT_USERS)
+      dialogModel: new DialogWithTwoButtonModel()
     };
+    this.store$.dispatch(prepareSelectionPopup({ selectedUsersIds: this.item.Users.map((user) => user.id) }));
 
     this.dialogService
       .open(UsersDialogComponent, dialogDataModel, DialogType.MAX)
       .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<User[]>) => {
-        if (result && result.success) {
-          const selectedUsers: User[] = result.model.filter(
-            (selectedUser) => !this.item.Users.some((user) => user.id === selectedUser.id)
+      .subscribe((result: IDialogResult<number[]>) => {
+        if (result?.success) {
+          const selectedUsersIds: number[] = result.data?.filter(
+            (selectedUserId) => !this.item.Users.some((user) => user.id === selectedUserId)
           );
-          if (selectedUsers?.length) {
-            this.item = {
-              ...this.item,
-              Users: [...this.item.Users, ...selectedUsers]
-            };
-            this.cdr.detectChanges();
+          if (selectedUsersIds?.length) {
+            this.store$.pipe(select(selectUsersById(selectedUsersIds)), first()).subscribe((selectedUsers) => {
+              this.item = {
+                ...this.item,
+                Users: [...this.item.Users, ...selectedUsers]
+              };
+              this.cdr.detectChanges();
+            });
           }
         }
+        this.store$.dispatch(resetSelectionPopup());
       });
-
-    // TODO: @ArsenIrod add afterOpened logic
   }
 
   addPrivilegeDialog(): void {
@@ -85,10 +87,9 @@ export class TemplatesRoleViewComponent extends TemplatesViewDetailsComponent<Ro
     this.dialogService
       .open(PrivilegesDialogComponent, dialogDataModel, DialogType.STANDART)
       .afterClosed()
-      .pipe(takeUntil(this.unsubscribe))
-      .subscribe((result: DialogResultModel<Privilege[]>) => {
-        if (result && result.success) {
-          const selectedPrivileges: Privilege[] = result.model
+      .subscribe((result: IDialogResult<Privilege[]>) => {
+        if (result?.success) {
+          const selectedPrivileges: Privilege[] = result.data
             .filter(
               (selectedPrivilege) => !this.item.Privileges.some((privilege) => privilege.id === selectedPrivilege.id)
             )
