@@ -14,14 +14,15 @@ import { MatPaginator } from '@angular/material/paginator';
 import { MatTable } from '@angular/material/table';
 import { CdkTable } from '@angular/cdk/table';
 import { CdkDrag, CdkDragDrop, CdkDropList, moveItemInArray } from '@angular/cdk/drag-drop';
-import { merge, Observable, Subject } from 'rxjs';
-import { map, takeUntil, tap } from 'rxjs/operators';
+import { ActivatedRoute, Params, Router } from '@angular/router';
+import { combineLatest, merge, Observable, Subject } from 'rxjs';
+import { delay, map, startWith, takeUntil, tap } from 'rxjs/operators';
 
 import { select, Store } from '@ngrx/store';
 
 import { IconsService } from '@/core/services';
 import { AppState } from '@/core/store';
-import { getItemsPerPageState } from '@/core/store/preferences';
+import { getDefaultListOutlineBorders, getItemsPerPageState } from '@/core/store/preferences';
 import {
   removeTableConfig,
   setTableConfig,
@@ -29,8 +30,6 @@ import {
   tableColumnsToDisplay,
   tableOutlineBorders
 } from '@/core/modules/layout/store';
-import { DataColumn } from '@/shared/models/table/data-column.model';
-import { DataRow } from '@/shared/models/table/data-row';
 import {
   ACTION_LABELS,
   BS_ICONS,
@@ -44,9 +43,11 @@ import {
   STYLECONSTANTS,
   THEME_PALETTE
 } from '@/shared/constants';
-import { HorizontalAlign } from '@/shared/models/table/horizontalAlign.enum';
 import {
   CellType,
+  DataRow,
+  HorizontalAlign,
+  IColumn,
   RowAction,
   RowActionData,
   RowActionType,
@@ -55,6 +56,8 @@ import {
 } from '@/shared/models/table';
 import { CommonDataSource } from '@/shared/services';
 import { PageQuery } from '@/shared/models';
+import { SelectionModel } from '@angular/cdk/collections';
+import { ListDisplayMode } from '@/shared/store';
 
 @Component({
   selector: 'table-component',
@@ -63,15 +66,18 @@ import { PageQuery } from '@/shared/models';
 })
 export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
   itemsPerPageState$: Observable<IItemsPerPage> = this.store$.pipe(select(getItemsPerPageState));
-  columnsToDisplay$: Observable<string[]>;
+  outlineBordersDefaultPreference$: Observable<boolean> = this.store$.pipe(select(getDefaultListOutlineBorders));
   outlineBorders$: Observable<boolean>;
+  columnsToDisplay$: Observable<string[]>;
 
   @Input() id: string;
   @Input() dataSource: CommonDataSource<DataRow>;
   @Input() totalItems: number;
-  @Input() columns: DataColumn[];
+  @Input() preselectedItems: number[];
+  @Input() columns: IColumn[];
   @Input() canSort = true;
   @Input() hasSettings = true;
+  @Input() displayMode = ListDisplayMode.DEFAULT;
   @Input() noContentMessage = CONSTANTS.NO_CONTENT_INFO;
   @Input() additionalRowActions: RowAction<RowActionType>[];
   @Input() canEdit: boolean;
@@ -85,13 +91,15 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
   @ViewChild(MatSort) sort: MatSort;
   @ViewChild(MatPaginator) paginator: MatPaginator;
 
+  selection: SelectionModel<number>;
+
   pageSizeOptions: number[] = pageSizeOptions;
   cellType: typeof CellType = CellType;
   buttonType: typeof BUTTON_TYPE = BUTTON_TYPE;
   matButtonType: typeof MAT_BUTTON = MAT_BUTTON;
   themePalette: typeof THEME_PALETTE = THEME_PALETTE;
   actionLabels: typeof ACTION_LABELS = ACTION_LABELS;
-  columnActionsKey = COLUMN_KEYS.ACTIONS;
+  columnKeys: typeof COLUMN_KEYS = COLUMN_KEYS;
   columnsInitialState: TableColumnConfig[];
 
   icons: { [key: string]: BS_ICONS } = {
@@ -136,7 +144,12 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
 
   private unsubscribe: Subject<void> = new Subject();
 
-  constructor(private readonly store$: Store<AppState>, private readonly iconsService: IconsService) {
+  constructor(
+    private readonly store$: Store<AppState>,
+    private readonly iconsService: IconsService,
+    private readonly router: Router,
+    private readonly activatedRoute: ActivatedRoute
+  ) {
     this.iconsService.registerIcons([...Object.values(this.icons)]);
   }
 
@@ -147,15 +160,38 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
     if (changes.additionalRowActions && this.additionalRowActions) {
       this.rowActions = [...this.rowActions, ...this.additionalRowActions];
     }
+    if (changes.preselectedItems && changes.displayMode && this.displayMode && this.preselectedItems) {
+      this.selection = new SelectionModel(
+        this.displayMode === ListDisplayMode.POPUP_MULTI_SELECTION,
+        this.preselectedItems
+      );
+    }
   }
 
   ngAfterViewInit(): void {
+    const listQueryParams: Partial<PageQuery> = this.validateListQueryParams(this.activatedRoute.snapshot.queryParams);
+    if (listQueryParams?.pageIndex) {
+      this.paginator.pageIndex = listQueryParams.pageIndex;
+    }
+    if (listQueryParams?.pageSize) {
+      this.paginator.pageSize = listQueryParams.pageSize;
+    }
+    if (listQueryParams?.sortIndex) {
+      this.sort.sort({
+        id: listQueryParams.sortIndex,
+        start: listQueryParams.sortDirection || 'asc',
+        disableClear: false
+      });
+    }
+
     const sort$ = this.sort.sortChange.pipe(tap(() => (this.paginator.pageIndex = 0)));
     merge(sort$, this.paginator.page)
-      .pipe(tap(() => this.loadDataPage()))
+      .pipe(
+        startWith([null, null]),
+        delay(0),
+        tap(() => this.loadDataPage())
+      )
       .subscribe();
-
-    this.loadDataPage();
   }
 
   ngOnDestroy(): void {
@@ -199,11 +235,11 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
   resetTableConfig(): void {
     this.columns = this.columns
       .sort(
-        (a: DataColumn, b: DataColumn) =>
+        (a: IColumn, b: IColumn) =>
           this.columnsInitialState.findIndex((col: TableColumnConfig) => col.title === a.key) -
           this.columnsInitialState.findIndex((col: TableColumnConfig) => col.title === b.key)
       )
-      .map((col: DataColumn, i: number) => ({ ...col, isVisible: this.columnsInitialState[i].isVisible }));
+      .map((col: IColumn, i: number) => ({ ...col, isVisible: this.columnsInitialState[i].isVisible }));
     this.store$.dispatch(removeTableConfig({ key: this.id }));
   }
 
@@ -224,16 +260,19 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
   }
 
   rowSelect(id: number): void {
-    // TODO: this is default action - add custom action on row select for example on modals
-    this.rowDedicatedAction(id, { actionType: RowActionType.DETAILS });
+    if (this.displayModePopup()) {
+      this.selectionChange(id);
+    } else {
+      this.rowDedicatedAction(id, { actionType: RowActionType.DETAILS });
+    }
   }
 
-  dropColumns(event: CdkDragDrop<DataColumn[]>): void {
+  dropColumns(event: CdkDragDrop<IColumn[]>): void {
     moveItemInArray(this.columns, event.previousIndex, event.currentIndex);
     this.updateTableConfig();
   }
 
-  dropColumnsPredicate(index: number, _: CdkDrag<DataColumn>, dropList: CdkDropList<DataColumn[]>) {
+  dropColumnsPredicate(index: number, _: CdkDrag<IColumn>, dropList: CdkDropList<IColumn[]>) {
     return dropList.data[index].draggable;
   }
 
@@ -252,23 +291,67 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
     }
   }
 
+  selectionChange(id: number): void {
+    this.selection.toggle(id);
+    this.rowActionClicked.emit({
+      actionType: RowActionType.SELECT,
+      ids: this.selection.selected
+    });
+  }
+
+  isHighlighted(id: number): boolean {
+    return this.displayMode === ListDisplayMode.POPUP_SINGLE_SELECTION && this.selection.isSelected(id);
+  }
+
+  displayModePopup(): boolean {
+    return (
+      this.displayMode === ListDisplayMode.POPUP_MULTI_SELECTION ||
+      this.displayMode === ListDisplayMode.POPUP_SINGLE_SELECTION
+    );
+  }
+
   private setColumns(): void {
     this.columnsInitialState = this.columns.map((col) => ({ title: col.key, isVisible: col.isVisible }));
-    this.outlineBorders$ = this.store$.pipe(select(tableOutlineBorders(this.id)));
+    this.outlineBorders$ = combineLatest([
+      this.outlineBordersDefaultPreference$,
+      this.store$.pipe(select(tableOutlineBorders(this.id)))
+    ]).pipe(
+      map(([defaultPreference, listPreference]) => (listPreference !== undefined ? listPreference : defaultPreference))
+    );
     this.columnsToDisplay$ = this.store$.pipe(
       select(tableColumnsToDisplay(this.id)),
-      map((columns) => (!columns ? this.columnsInitialState.filter((c) => c.isVisible).map((c) => c.title) : columns))
+      map((columns) => (!columns ? this.columnsInitialState.filter((c) => c.isVisible).map((c) => c.title) : columns)),
+      map((columnsToDisplay) => {
+        if (this.displayModePopup()) {
+          if (this.displayMode === ListDisplayMode.POPUP_MULTI_SELECTION) {
+            const selectIndex = this.columns.findIndex((col) => col.key === COLUMN_KEYS.SELECT);
+            if (selectIndex >= 0) {
+              columnsToDisplay = Object.assign([], columnsToDisplay, { [selectIndex]: COLUMN_KEYS.SELECT });
+            }
+          }
+          return columnsToDisplay.filter((cTitle) => cTitle !== COLUMN_KEYS.ACTIONS);
+        }
+        return columnsToDisplay;
+      })
     );
     this.store$.pipe(select(tableColumnsConfig(this.id)), takeUntil(this.unsubscribe)).subscribe((columnsConfig) => {
       if (columnsConfig) {
         this.columns = this.columns
           .sort(
-            (a: DataColumn, b: DataColumn) =>
+            (a: IColumn, b: IColumn) =>
               columnsConfig.findIndex((col: TableColumnConfig) => col.title === a.key) -
               columnsConfig.findIndex((col: TableColumnConfig) => col.title === b.key)
           )
-          .map((col: DataColumn, i: number) => ({ ...col, isVisible: columnsConfig[i].isVisible }));
+          .map((col: IColumn, i: number) => ({ ...col, isVisible: columnsConfig[i].isVisible }));
       }
+    });
+  }
+
+  private setQueryParams(queryParams: Params): void {
+    this.router.navigate([], {
+      relativeTo: this.activatedRoute,
+      queryParams: queryParams,
+      queryParamsHandling: 'merge'
     });
   }
 
@@ -279,7 +362,30 @@ export class TableComponent implements OnChanges, AfterViewInit, OnDestroy {
       sortIndex: this.sort.active || COLUMN_KEYS.ID,
       sortDirection: this.sort.direction || SORT_DIRECTION.ASC
     };
-
     this.dataSource.loadData(newPage);
+    if (!this.displayModePopup()) {
+      this.setQueryParams(newPage);
+    }
+  }
+
+  private validateListQueryParams(params: Params): Partial<PageQuery> {
+    let pageQuery: Partial<PageQuery> = {};
+    if (params.pageSize) {
+      const pageSize = parseInt(params.pageSize, 0);
+      if (this.pageSizeOptions.includes(pageSize)) {
+        pageQuery = { ...pageQuery, pageSize };
+      }
+    }
+    if (params.pageIndex) {
+      const pageIndex = parseInt(params.pageIndex, 0);
+      pageQuery = { ...pageQuery, pageIndex };
+    }
+    if (params.sortIndex && this.columns.some((col) => col.key === params.sortIndex && col.hasSorting)) {
+      pageQuery = { ...pageQuery, sortIndex: params.sortIndex };
+    }
+    if (params.sortDirection && ['asc', 'desc'].includes(params.sortDirection)) {
+      pageQuery = { ...pageQuery, sortDirection: params.sortDirection };
+    }
+    return pageQuery;
   }
 }
