@@ -1,5 +1,6 @@
 import { Service } from 'typedi';
 import { Strategy, ExtractJwt, StrategyOptions } from 'passport-jwt';
+import { BasicStrategy } from 'passport-http';
 import passport from 'passport';
 
 import { Config } from './config';
@@ -10,17 +11,22 @@ export class Passport {
   constructor(private readonly userService: UserService) {}
 
   private opts: StrategyOptions = {
-    jwtFromRequest: ExtractJwt.fromAuthHeaderWithScheme('JWT'),
+    jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
     secretOrKey: process.env.ACCESS_TOKEN_SECRET,
     // issuer: 'auth@mywebmaster.pp.ua',
     audience: Config.WEB_URL
   };
 
-  public authenticate() {
-    return passport.authenticate('jwt', {
-      session: false,
-      failWithError: true
-    });
+  public authenticate(strategy: 'jwt' | 'basic' = 'jwt') {
+    switch (strategy) {
+      case 'jwt':
+        return passport.authenticate('jwt', {
+          session: false,
+          failWithError: true
+        });
+      case 'basic':
+        return passport.authenticate('basic', { session: false });
+    }
   }
 
   init(): void {
@@ -37,6 +43,14 @@ export class Passport {
           }
         } catch (error) {
           return done(error, false);
+        }
+      })
+    );
+
+    passport.use(
+      new BasicStrategy((username, password, done) => {
+        if (username === Config.OPENAPI_USERNAME && password === Config.OPENAPI_PASSWORD) {
+          return done(null, username);
         }
       })
     );
