@@ -1,7 +1,7 @@
 import { Response, Router } from 'express';
 import { Service } from 'typedi';
 
-import { Preference, RequestWithBody, BaseResponse, PreferenceCreationAttributes } from '../models';
+import { Preference, RequestWithBody, BaseResponse, PreferenceCreationAttributes, ItemApiResponse } from '../models';
 import { PreferenceController } from '../controllers';
 
 @Service()
@@ -15,8 +15,10 @@ export class PreferenceRoutes {
 
     this.router.post(
       '/',
-      async (req: RequestWithBody<PreferenceCreationAttributes>, res: Response<Preference | BaseResponse>) =>
-        this.preferenceController.set(req, res)
+      async (
+        req: RequestWithBody<PreferenceCreationAttributes>,
+        res: Response<ItemApiResponse<Preference> | BaseResponse>
+      ) => this.preferenceController.set(req, res)
     );
 
     return this.router;
