@@ -6,6 +6,7 @@ import { provideMockStore } from '@ngrx/store/testing';
 
 import { SharedModule } from '@/shared/shared.module';
 import { routerStoreMock } from '@/shared/testing/mocks';
+import { RoutingConstants } from '@/shared/constants';
 import { Breadcrumb } from './breadcrumb';
 import { BreadcrumbsComponent } from './breadcrumbs.component';
 
@@ -41,7 +42,7 @@ describe('BreadcrumbsComponent', () => {
       expect(breadcrumbs).toEqual([
         {
           keyString: 'home',
-          url: '/home'
+          url: RoutingConstants.ROUTE_HOME
         },
         {
           keyString: 'Planner',
