@@ -8,7 +8,7 @@ import { PageQuery } from '@/shared/models';
 import { CommonDataSource } from '@/shared/services';
 import { COLUMN_KEYS } from '@/shared/constants/table.constants';
 import { UrlGenerator } from '@/shared/utils/url.generator';
-import { DataRow, CellValue } from '@/shared/models/table';
+import { DataRow, Cell } from '@/shared/models/table';
 import { selectDepartmentsOfPage } from '../store/department.selectors';
 
 export class DepartmentsDataSource extends CommonDataSource<Department> {
@@ -31,16 +31,16 @@ export class DepartmentsDataSource extends CommonDataSource<Department> {
   protected mapToDataRows(departments: Department[]): DataRow[] {
     return departments.map((department) => ({
       id: department.id,
-      [COLUMN_KEYS.SEQUENCE]: CellValue.createSequenceCell(),
-      [COLUMN_KEYS.TITLE]: CellValue.createStringCell(department.title),
-      [COLUMN_KEYS.MANAGER]: CellValue.createLinkCell(
+      [COLUMN_KEYS.SEQUENCE]: Cell.createSequenceCell(),
+      [COLUMN_KEYS.TITLE]: Cell.createStringCell(department.title),
+      [COLUMN_KEYS.MANAGER]: Cell.createLinkCell(
         department.Manager?.fullname,
         UrlGenerator.getUserUrl(department.Manager?.id)
       ),
-      [COLUMN_KEYS.WORKERS]: CellValue.createStringCell(department.Workers?.length),
-      [COLUMN_KEYS.CREATED_AT]: CellValue.createDateCell(department.createdAt),
-      [COLUMN_KEYS.UPDATED_AT]: CellValue.createDateCell(department.updatedAt),
-      [COLUMN_KEYS.ACTIONS]: CellValue.createActionsCell()
+      [COLUMN_KEYS.WORKERS]: Cell.createStringCell(department.Workers?.length),
+      [COLUMN_KEYS.CREATED_AT]: Cell.createDateCell(department.createdAt),
+      [COLUMN_KEYS.UPDATED_AT]: Cell.createDateCell(department.updatedAt),
+      [COLUMN_KEYS.ACTIONS]: Cell.createActionsCell()
     }));
   }
 }
