@@ -7,7 +7,7 @@ import { listPageRequested } from '@/core/modules/plan-api/store/plan';
 import { PageQuery } from '@/shared/models';
 import { CommonDataSource } from '@/shared/services';
 import { COLUMN_KEYS } from '@/shared/constants';
-import { DataRow, CellValue } from '@/shared/models/table';
+import { DataRow, Cell } from '@/shared/models/table';
 import { UrlGenerator } from '@/shared/utils';
 import { selectPlansOfPage } from '../store';
 
@@ -31,17 +31,14 @@ export class PlansDataSource extends CommonDataSource<Plan> {
   protected mapToDataRows(plans: Plan[]): DataRow[] {
     return plans.map((plan) => ({
       id: plan.id,
-      [COLUMN_KEYS.SEQUENCE]: CellValue.createSequenceCell(),
-      [COLUMN_KEYS.TITLE]: CellValue.createStringCell(plan.title),
-      [COLUMN_KEYS.STAGE]: CellValue.createStringCell(plan.activeStage?.keyString),
-      [COLUMN_KEYS.CREATOR]: CellValue.createLinkCell(
-        plan.Creator?.fullname,
-        UrlGenerator.getUserUrl(plan.Creator?.id)
-      ),
-      [COLUMN_KEYS.PARTICIPANTS]: CellValue.createStringCell(plan.Participants?.length),
-      [COLUMN_KEYS.CREATED_AT]: CellValue.createDateCell(plan.createdAt),
-      [COLUMN_KEYS.UPDATED_AT]: CellValue.createDateCell(plan.updatedAt),
-      [COLUMN_KEYS.ACTIONS]: CellValue.createActionsCell()
+      [COLUMN_KEYS.SEQUENCE]: Cell.createSequenceCell(),
+      [COLUMN_KEYS.TITLE]: Cell.createStringCell(plan.title),
+      [COLUMN_KEYS.STAGE]: Cell.createStringCell(plan.activeStage?.keyString),
+      [COLUMN_KEYS.CREATOR]: Cell.createLinkCell(plan.Creator?.fullname, UrlGenerator.getUserUrl(plan.Creator?.id)),
+      [COLUMN_KEYS.PARTICIPANTS]: Cell.createStringCell(plan.Participants?.length),
+      [COLUMN_KEYS.CREATED_AT]: Cell.createDateCell(plan.createdAt),
+      [COLUMN_KEYS.UPDATED_AT]: Cell.createDateCell(plan.updatedAt),
+      [COLUMN_KEYS.ACTIONS]: Cell.createActionsCell()
     }));
   }
 }
