@@ -6,7 +6,7 @@ import { listPageRequested } from '@/core/modules/role-api/store/role';
 import { Role } from '@/core/modules/role-api/shared';
 import { PageQuery } from '@/shared/models';
 import { CommonDataSource } from '@/shared/services';
-import { CellValue, DataRow } from '@/shared/models/table';
+import { Cell, DataRow } from '@/shared/models/table';
 import { selectRolesOfPage } from '../store/role.selectors';
 import { COLUMN_KEYS } from '@/shared/constants/table.constants';
 
@@ -30,13 +30,13 @@ export class RolesDataSource extends CommonDataSource<Role> {
   protected mapToDataRows(roles: Role[]): DataRow[] {
     return roles.map((role) => ({
       id: role.id,
-      [COLUMN_KEYS.SEQUENCE]: CellValue.createSequenceCell(),
-      [COLUMN_KEYS.TITLE]: CellValue.createStringCell(role.keyString),
-      [COLUMN_KEYS.USERS]: CellValue.createStringCell(role.Users?.length),
-      [COLUMN_KEYS.PRIVILEGES]: CellValue.createStringCell(role.Privileges?.length),
-      [COLUMN_KEYS.CREATED_AT]: CellValue.createDateCell(role.createdAt),
-      [COLUMN_KEYS.UPDATED_AT]: CellValue.createDateCell(role.updatedAt),
-      [COLUMN_KEYS.ACTIONS]: CellValue.createActionsCell()
+      [COLUMN_KEYS.SEQUENCE]: Cell.createSequenceCell(),
+      [COLUMN_KEYS.TITLE]: Cell.createStringCell(role.keyString),
+      [COLUMN_KEYS.USERS]: Cell.createStringCell(role.Users?.length),
+      [COLUMN_KEYS.PRIVILEGES]: Cell.createStringCell(role.Privileges?.length),
+      [COLUMN_KEYS.CREATED_AT]: Cell.createDateCell(role.createdAt),
+      [COLUMN_KEYS.UPDATED_AT]: Cell.createDateCell(role.updatedAt),
+      [COLUMN_KEYS.ACTIONS]: Cell.createActionsCell()
     }));
   }
 }
