@@ -1,6 +1,7 @@
 import { Action, createReducer, on } from '@ngrx/store';
 
 import * as roleApiActions from '@/core/modules/role-api/store/role/role.actions';
+import { Role } from '@/core/modules/role-api/shared';
 import { initialListState, ListState, pagesAdapter } from '@/shared/store';
 import * as roleActions from './role.actions';
 
@@ -8,7 +9,14 @@ const rolesReducer = createReducer(
   initialListState,
   on(roleActions.changeIsEditingState, (state, { isEditing }) => ({
     ...state,
-    editing: isEditing
+    isEditing
+  })),
+  on(roleActions.roleCached, (state, { displayedItemCopy }) => ({
+    ...state,
+    cache: {
+      ...state.cache,
+      displayedItemCopy
+    }
   })),
   on(roleApiActions.listPageRequested, (state) => ({
     ...state,
@@ -36,7 +44,7 @@ const rolesReducer = createReducer(
   })),
   on(roleApiActions.updateRoleSuccess, (state) => ({
     ...state,
-    editing: false
+    isEditing: false
   })),
   on(roleApiActions.rolesApiError, (state) => ({
     ...state,
@@ -44,7 +52,7 @@ const rolesReducer = createReducer(
   }))
 );
 
-export function reducer(state: ListState | undefined, action: Action) {
+export function reducer(state: ListState<Role> | undefined, action: Action) {
   return rolesReducer(state, action);
 }
 
