@@ -11,7 +11,7 @@ import { ListState, Page } from '@/shared/store';
 import { generatePageKey } from '@/shared/utils/generatePageKey';
 import * as fromRole from './role.reducer';
 
-export const selectRolesState = createFeatureSelector<ListState>(fromRole.rolesFeatureKey);
+export const selectRolesState = createFeatureSelector<ListState<Role>>(fromRole.rolesFeatureKey);
 export const selectRolePagesState = createSelector(selectRolesState, (rolesState) => rolesState?.pages);
 
 export const selectRolePageByKey = (pageQuery: PageQuery) =>
@@ -33,4 +33,6 @@ export const selectRolesOfPage = (pageQuery: PageQuery) =>
     }
   );
 
-export const selectIsEditing = createSelector(selectRolesState, (roleState) => roleState?.editing);
+export const selectIsEditing = createSelector(selectRolesState, (roleState) => roleState?.isEditing);
+
+export const selectRoleFromCache = createSelector(selectRolesState, (rolesState) => rolesState.cache.displayedItemCopy);
