@@ -4,8 +4,6 @@ import { Observable } from 'rxjs';
 
 import { Store, select } from '@ngrx/store';
 
-import { cloneDeep } from 'lodash';
-
 import { environment } from 'environments/environment';
 import { AppState } from '@/core/store';
 import { Preferences } from '@/core/store/preferences';
@@ -45,7 +43,6 @@ export class TemplatesUserProfileComponent implements OnInit, OnChanges {
   baseUrl = environment.baseUrl;
   coverUrl = CONSTANTS.NO_IMAGE_URL;
   coverTitle = CONSTANTS.NO_IMAGE_TITLE;
-  userInitial: User;
 
   constructor(private store: Store<AppState>, private route: ActivatedRoute) {}
 
@@ -61,7 +58,6 @@ export class TemplatesUserProfileComponent implements OnInit, OnChanges {
 
   ngOnChanges(changes: SimpleChanges): void {
     if (changes['user']?.currentValue && this.user) {
-      this.userInitial = cloneDeep(this.user);
       if (this.user.avatar) {
         this.setCover(this.user.avatar);
       }
@@ -73,21 +69,16 @@ export class TemplatesUserProfileComponent implements OnInit, OnChanges {
     this.coverTitle = cover.title;
   }
 
-  updateUserStore(asset?: Asset): void {
+  updateUser(user: User, asset?: Asset): void {
     if (asset) {
       this.setCover(asset);
+      user = { ...user, ...(asset && { avatar: asset, avatarId: asset.id }) };
     }
-    const user = cloneDeep({ ...this.user, ...(asset && { avatar: asset, avatarId: asset.id }) });
     if (this.isProfilePage) {
       this.store.dispatch(updateUserProfileRequested({ user }));
     } else {
       this.store.dispatch(updateUserRequested({ user }));
     }
-  }
-
-  updateUser(user: User): void {
-    this.user = user;
-    this.updateUserStore();
   }
 
   updateUserOrg(organization: Organization): void {
