@@ -1,14 +1,29 @@
 import { Action, on, createReducer } from '@ngrx/store';
 
 import * as userApiActions from '@/core/modules/user-api/store/user-api.actions';
-import { initialListState, pagesAdapter, ListState } from '@/shared/store';
+import { initialListState, pagesAdapter, ListState, ListDisplayMode } from '@/shared/store';
+import { User } from '@/core/modules/user-api/shared';
 import * as userActions from './user.actions';
+
+export interface UserListState extends ListState<User> {
+  selectedUsersIds: number[] | null;
+}
 
 const reducer = createReducer(
   initialListState,
   on(userActions.changeIsEditingState, (state, { isEditing }) => ({
     ...state,
-    editing: isEditing
+    isEditing
+  })),
+  on(userActions.prepareSelectionPopup, (state, { selectedUsersIds, singleSelection }) => ({
+    ...state,
+    listDisplayMode: singleSelection ? ListDisplayMode.POPUP_SINGLE_SELECTION : ListDisplayMode.POPUP_MULTI_SELECTION,
+    selectedUsersIds
+  })),
+  on(userActions.resetSelectionPopup, (state) => ({
+    ...state,
+    listDisplayMode: ListDisplayMode.DEFAULT,
+    selectedUsersIds: null
   })),
   on(userApiActions.listPageRequested, (state) => ({ ...state, pages: { ...state.pages, pageLoading: true } })),
   on(userApiActions.listPageLoaded, (state, { page, response: { pages, resultsNum } }) => ({
@@ -22,7 +37,7 @@ const reducer = createReducer(
   })),
   on(userApiActions.updateUserSuccess, (state) => ({
     ...state,
-    editing: false
+    isEditing: false
   })),
   on(userApiActions.deleteUser, (state) => ({
     ...state,
@@ -38,7 +53,7 @@ const reducer = createReducer(
   }))
 );
 
-export function usersReducer(state: ListState | undefined, action: Action) {
+export function usersReducer(state: UserListState | undefined, action: Action) {
   return reducer(state, action);
 }
 
