@@ -1,18 +1,20 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatIconTestingModule } from '@angular/material/icon/testing';
+import { RouterTestingModule } from '@angular/router/testing';
 
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 import { Store } from '@ngrx/store';
 
 import { initialPreferencesState } from '@/core/store/preferences';
 import { AppState } from '@/core/store';
-import { CellValue, DataColumn, DataRow } from '@/shared/models/table';
+import { initialLayoutState } from '@/core/modules/layout/store';
+import { Cell, Column, DataRow, IColumn } from '@/shared/models/table';
 import { SharedModule } from '@/shared/shared.module';
 import { COLUMN_KEYS } from '@/shared/constants';
 import { CommonDataSource } from '@/shared/services';
 import { PageQuery } from '@/shared/models';
 import { TableComponent } from './table.component';
-import { MatIconTestingModule } from '@angular/material/icon/testing';
 
 interface TestData {
   id: number;
@@ -37,11 +39,11 @@ class TestDataSource extends CommonDataSource<TestData> {
   protected mapToDataRows(items: any[]): DataRow[] {
     return items.map((item: { id: any; title: string; createdAt: Date; updatedAt: Date }) => ({
       id: item.id,
-      [COLUMN_KEYS.SEQUENCE]: CellValue.createSequenceCell(),
-      [COLUMN_KEYS.TITLE]: CellValue.createStringCell(item.title),
-      [COLUMN_KEYS.CREATED_AT]: CellValue.createDateCell(item.createdAt),
-      [COLUMN_KEYS.UPDATED_AT]: CellValue.createDateCell(item.updatedAt),
-      [COLUMN_KEYS.ACTIONS]: CellValue.createActionsCell()
+      [COLUMN_KEYS.SEQUENCE]: Cell.createSequenceCell(),
+      [COLUMN_KEYS.TITLE]: Cell.createStringCell(item.title),
+      [COLUMN_KEYS.CREATED_AT]: Cell.createDateCell(item.createdAt),
+      [COLUMN_KEYS.UPDATED_AT]: Cell.createDateCell(item.updatedAt),
+      [COLUMN_KEYS.ACTIONS]: Cell.createActionsCell()
     }));
   }
 }
@@ -50,21 +52,23 @@ describe('TableComponent', () => {
   let component: TableComponent;
   let fixture: ComponentFixture<TableComponent>;
   const initialState = {
-    preferences: initialPreferencesState
+    preferences: initialPreferencesState,
+    layout: initialLayoutState
   };
   let store: MockStore;
-  const columns = [
-    DataColumn.createSequenceNumberColumn(),
-    DataColumn.createColumn({ key: COLUMN_KEYS.TITLE }),
-    DataColumn.createColumn({ key: COLUMN_KEYS.CREATED_AT }),
-    DataColumn.createColumn({ key: COLUMN_KEYS.UPDATED_AT }),
-    DataColumn.createActionsColumn()
+  const columns: IColumn[] = [
+    Column.createSequenceNumberColumn(),
+    Column.createCheckboxColumn(),
+    Column.createColumn({ key: COLUMN_KEYS.TITLE }),
+    Column.createColumn({ key: COLUMN_KEYS.CREATED_AT }),
+    Column.createColumn({ key: COLUMN_KEYS.UPDATED_AT }),
+    Column.createActionsColumn()
   ];
 
   beforeEach(
     waitForAsync(() => {
       TestBed.configureTestingModule({
-        imports: [SharedModule, MatIconTestingModule, BrowserAnimationsModule],
+        imports: [SharedModule, MatIconTestingModule, RouterTestingModule, BrowserAnimationsModule],
         declarations: [TableComponent],
         providers: [provideMockStore({ initialState })]
       }).compileComponents();
@@ -77,6 +81,8 @@ describe('TableComponent', () => {
     component = fixture.componentInstance;
     component.columns = columns;
     component.dataSource = new TestDataSource(store as Store<AppState>);
+    component.totalItems = 1;
+    component.id = 'test-list';
     fixture.detectChanges();
   });
 
