@@ -2,6 +2,6 @@ export * from './dialog-with-two-button.model';
 export * from './dialog-create-edit.model';
 export * from './dialog-mode.enum';
 export * from './dialog-data.model';
-export * from './dialog-result.model';
+export * from './dialog-result';
 export * from './dialog-type.enum';
 export * from './dialog-confirm.model';
