@@ -1,6 +1,6 @@
-import { CellValue } from './cell-value.model';
+import { ICell } from './cell.model';
 
 export interface DataRow {
   id: number;
-  [key: string]: CellValue | number;
+  [key: string]: ICell | number;
 }
