@@ -1,6 +1,6 @@
-export * from './data-column.model';
+export * from './column.model';
 export * from './data-row';
-export * from './cell-value.model';
+export * from './cell.model';
 export * from './horizontalAlign.enum';
 export * from './cellType.enum';
 export * from './row-action';
