@@ -1,5 +1,11 @@
 import { createEntityAdapter, EntityAdapter, EntityState } from '@ngrx/entity';
 
+export enum ListDisplayMode {
+  DEFAULT,
+  POPUP_MULTI_SELECTION,
+  POPUP_SINGLE_SELECTION
+}
+
 export interface Page {
   key: string;
   dataIds: number[];
@@ -9,6 +15,10 @@ export interface PagesState extends EntityState<Page> {
   pageLoading: boolean;
   pages: number;
   resultsNum: number;
+}
+
+export interface CacheState<T> {
+  displayedItemCopy: T;
 }
 
 export const pagesAdapter: EntityAdapter<Page> = createEntityAdapter<Page>({
@@ -22,14 +32,22 @@ export const initialPagesState: PagesState = pagesAdapter.getInitialState({
   resultsNum: null
 });
 
-export interface ListState {
-  editing: boolean;
+export const initialCacheState: CacheState<unknown> = {
+  displayedItemCopy: null
+};
+
+export interface ListState<T> {
+  isEditing: boolean;
   pages: PagesState;
+  listDisplayMode: ListDisplayMode;
+  cache: CacheState<T>;
 }
 
-export const initialListState = {
-  editing: false,
-  pages: initialPagesState
+export const initialListState: ListState<unknown> = {
+  isEditing: false,
+  pages: initialPagesState,
+  listDisplayMode: ListDisplayMode.DEFAULT,
+  cache: initialCacheState
 };
 
 export const {
